+++ CHANGELOG.md
@@ -3699,3 +3699,4 @@
 - [2026-08-18 13:37] feat(test): add bi-directional event bus handler for test
 - [2026-08-18 13:45] feat(test): implement shimmer loading skeleton in test
 - [2026-08-18 13:47] feat(i18n): add async i18n stream provider
+- [2026-08-18 14:11] feat(storage): integrate custom telemetry logger in storage